Big improvement of collision system - collider struct, groundCollisionData, tileCollisionFlags, onTileCollidesWithPlayer event

+++ inc/hud.h
@@ -3,4 +3,4 @@
 void hud_init();
 void hud_update();
 void hud_updateSprites();
-void hud_setHP(u8 val);
+void hud_setHP(s16 val);
